(refactor & fix) Move elemental shaman maelstrom tab to own features folder.
